backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
alembic==1.18.4
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
pydantic[email]==2.12.5
pydantic-settings==2.13.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
httpx==0.28.1
